requirements.txt
faster-whisper==0.10.0
python-dotenv==1.0.1
